完成回收员生成账单功能

+++ yueyishou-parent/model/src/main/java/com/ilhaha/yueyishou/model/entity/coupon/CustomerCoupon.java
@@ -19,14 +19,14 @@ public class CustomerCoupon extends BaseEntity {
     private Long couponId;
 	/**顾客ID*/
     private Long customerId;
-	/**优化券状态（1：未使用 2：已使用）*/
+	/**服务抵扣劵券状态（1：未使用 2：已使用）*/
     private Integer status;
 	/**领取时间*/
     private Date receiveTime;
 	/**使用时间*/
     private Date usedTime;
 	/**使用的订单id*/
-    private Integer orderId;
+    private Long orderId;
 	/**过期时间*/
     private Date expireTime;
 }
